pysemgrep: extract new core_targets_plan.py from core_runner.py
core_runner.py was getting big too.

test plan:
make e2e
